deletion requires valid authentication

+++ controllers/blogs.js
@@ -57,6 +57,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+
+  if (!token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const target = await Blog.findById(request.params.id)
+
+  if (target.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'incorrect user'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
